Commented the hell out of the main cpp file & fixed the outstanding scale issue with map view. Map view can now extend up to lightyears in scale if necessary

+++ Ignition_Engine.h
@@ -18,7 +18,7 @@ class Ignition_engine
 	bool map_view;
 	long double deltat;	
 	double Redraw_interval;	
-	int zoom_exponent;
+	long long int zoom_exponent;
 	int time_acceleration_exponent;
 	long double simulation_time;
 	std::string Program_version;
